execution/protocol/rules/ethash: move ShouldReorg from headerdownload
The headerdownload package is slated for removal along with the legacy
stage_headers/bodies pipeline. Park ShouldReorg in ethash so the offline
import command can keep calling it after headerdownload is gone.
